changing vad api, adding pipeline example

+++ modules/lib_vad/api/vad_api.h
@@ -36,7 +36,7 @@ void vad_init(vad_state_t * state);
  *                             very likely not voice, and 255 indicates very
  *                             likely voice
  */
-int32_t vad_probability_voice(const int32_t input[VAD_FRAME_ADVANCE],
+uint8_t vad_probability_voice(const int32_t input[VAD_FRAME_ADVANCE],
                             vad_state_t * state);
 
 //void vad_test_task(chanend app_to_vad, chanend vad_to_app, chanend ?c_control);
